Refactor and gate contract tests
